+++ index.common.js
@@ -2,10 +2,6 @@
 
 exports = module.exports = require('./build/reactTimeAgo').default
 
-exports.TimeAgo             = require('./build/time ago').default
-// Legacy export
-exports.Time_ago            = require('./build/time ago').default
-
 exports.Date_time_formatter = require('./build/date time formatter').default
 
 exports['default'] = exports