+++ index.es6.js
@@ -1,5 +1,2 @@
 export { default as default } from './es6/ReactTimeAgo'
-export { default as TimeAgo } from './es6/time ago'
-// Legacy export
-export { default as Time_ago } from './es6/time ago'
 export { default as Date_time_formatter } from './es6/date time formatter'